+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   trigger function helper.
 * Remove `Read` instance of `IndexMethod`.
 * Allow using nonreserved keywords, e.g. `timestamp`, in table indices.
+* Add support for renaming an existing local index during a migration that
+  creates index concurrently.
 
 # hpqtypes-extras-1.19.0.0 (2025-11-27)
 * Compatibility with `hpqtypes` >= 0.13.0.0.